Fix heap corruption issue in DriverProxy
This commit fixes an issue where the buffer used by the DriverProxy class can become unpinned causing clients to send corrupt messages to the Media Driver.

UnsafeBuffer has a constructor that takes a ByteBuffer.  ByteBuffer contains a pointer to a managed array that has been pinned with a GCHandle.  When an UnsafeBuffer is constructed from a ByteBuffer it takes a copy of the pointer, but does not take ownership of the GCHandle.  If the ByteBuffer is allowed to be garbage collected the GCHandle is freed in its finalizer unpinning the managed array.  The unpinned array may be moved by the garbage collector leaving UnsafeBuffer pointing to invalid memory.

This commit fixes the bug by ensuring that we keep a reference to the ByteBuffer that is wrapped by the UnsafeBuffer in DriverProxy. This prevents the garbage collector from calling the ByteBuffer's finalizer and unpinning the managed array whilst it is still being used by the DriverProxy.  It also fixes a similar issue in the SimplePublisher sample.